1.增加岗位接口与企业接口 2.数据的一些校验

+++ src/main/java/com/zerody/user/pojo/SysLoginInfo.java
@@ -18,8 +18,8 @@ public class SysLoginInfo extends BaseStringModel {
     //手机号
     private String mobileNumber;
 
-    //是否激活
-    private Byte activeFlag;
+    //是否启用：类型：1启用，2企业停用 3部门停用
+    private Integer activeFlag;
 
     //用户ID
     private String userId;
@@ -44,6 +44,5 @@ public class SysLoginInfo extends BaseStringModel {
 
     //修改时间
     private Date updateTime;
-
 
 }
